// front-end-app/src/app/models/Internaute.model.ts
export class Internaute {
  _id?: string;
  email?: string;
  password?: string;
  currentPassword?: string;
  passwordConfirm?: string;
  plainPassword?: string;
  firstName?: string;
  lastName?: string;
  phoneNumber?: string;
  licenceNumber?: string;
  licencePlateNumber?: string;
  type?: string;
}